Move LoginManager and SessionManager to the Security namespace
